third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/undo-dom-edits-2-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that DOM modifications done in the Elements panel are undoable (Part 2).


Running: testSetUp
Setting up

Running: testSetAttribute
Initial:
- <div id="testSetAttribute">
      <div foo="attribute value" id="node-to-set-attribute"></div>
  </div>
Post-action:
- <div id="testSetAttribute">
      <div id="node-to-set-attribute" bar="edited attribute"></div>
  </div>
Post-undo (initial):
- <div id="testSetAttribute">
      <div id="node-to-set-attribute" foo="attribute value"></div>
  </div>
Post-redo (action):
- <div id="testSetAttribute">
      <div id="node-to-set-attribute" bar="edited attribute"></div>
  </div>

Running: testRemoveAttribute
Initial:
- <div id="testRemoveAttribute">
      <div foo="attribute value" id="node-to-remove-attribute"></div>
  </div>
Post-action:
- <div id="testRemoveAttribute">
      <div id="node-to-remove-attribute"></div>
  </div>
Post-undo (initial):
- <div id="testRemoveAttribute">
      <div id="node-to-remove-attribute" foo="attribute value"></div>
  </div>
Post-redo (action):
- <div id="testRemoveAttribute">
      <div id="node-to-remove-attribute"></div>
  </div>

Running: testAddAttribute
Initial:
- <div id="testAddAttribute">
      <div id="node-to-add-attribute"></div>
  </div>
Post-action:
- <div id="testAddAttribute">
      <div id="node-to-add-attribute" newattr="new-value"></div>
  </div>
Post-undo (initial):
- <div id="testAddAttribute">
      <div id="node-to-add-attribute"></div>
  </div>
Post-redo (action):
- <div id="testAddAttribute">
      <div id="node-to-add-attribute" newattr="new-value"></div>
  </div>
